fix(table): fix label display when a command + filter + label is provided
currently, if you use `:svc -l app.kubernetes.io/instance=cert-manager` in k9s and then navigate history, when returning to the page the results are correct but the filter is not shown. here, we ensure that the table title string is not an empty string.

reported in #3404, but unrelated

bug originally introduced in #3293, which I have not reviewed for any other rippling effects.